Implementing some of the changes requested in peer review
