+++ test/internal/modules/invoice-accounts/controllers/contacts.js
@@ -4,9 +4,7 @@ const {
   experiment,
   test,
   beforeEach,
-  afterEach,
-  before,
-  after
+  afterEach
 } = exports.lab = require('@hapi/lab').script();
 const controller = require('../../../../../src/internal/modules/invoice-accounts/controllers/contacts');
 const uuid = require('uuid');
@@ -254,72 +252,6 @@ experiment('./internal/modules/invoice-accounts/controller', () => {
         await controller.postContactCreate(request, h);
         const args = h.postRedirectGet.lastCall.args;
         expect(args[1]).to.equal(`/invoice-accounts/create/${regionId}/${companyId}/create-contact`);
-      });
-    });
-  });
-
-  experiment('.getContactSearch', () => {
-    beforeEach(async () => {
-      await dataService.sessionManager.returns(getSessionData());
-      await controller.getContactSearch(request, h);
-    });
-    afterEach(async () => {
-      await sandbox.restore();
-    });
-    test('calls sessionManager with the correct params', async () => {
-      const args = dataService.sessionManager.lastCall.args;
-      expect(args[0]).to.equal(request);
-      expect(args[1]).to.equal(regionId);
-      expect(args[2]).to.equal(companyId);
-      // no data to merge is passed to the session
-      expect(args[3]).to.equal(undefined);
-    });
-
-    test('the correct data is passed to the view', async () => {
-      const sessionData = getSessionData();
-      dataService.sessionManager.returns(sessionData);
-      await controller.getContactSearch(request, h);
-      const args = h.view.lastCall.args[1];
-      expect(args.back).to.startWith(`/invoice-accounts/create/${regionId}/${companyId}`);
-      expect(args.pageTitle).to.equal('Does this contact already exist?');
-    });
-  });
-
-  experiment('.postContactSearch', async () => {
-    experiment('when the form is valid', () => {
-      experiment('when the user opts to create a new contact', () => {
-        let modifiedRequest;
-        beforeEach(async () => {
-          modifiedRequest = { payload: { filter: 'some string', id: 'new' } };
-          Object.assign(modifiedRequest, request);
-          await dataService.sessionManager.returns(getSessionData());
-          await controller.postContactSearch(modifiedRequest, h);
-        });
-        afterEach(async () => {
-          await sandbox.restore();
-        });
-        test('then redirect to contact entry', async () => {
-          await controller.postContactSearch(modifiedRequest, h);
-          const args = h.redirect.lastCall.args;
-          expect(args[0]).to.startWith(`/contact-entry/new`);
-        });
-      });
-      experiment('when the user opts to select an existing contact', () => {
-        const contactId = uuid();
-        let modifiedRequest;
-        before(async () => {
-          modifiedRequest = { payload: { filter: 'some string', id: contactId } };
-          Object.assign(modifiedRequest, request);
-          await controller.postContactSearch(modifiedRequest, h);
-        });
-        after(async () => {
-          await sandbox.restore();
-        });
-        test('then redirect to select an address', async () => {
-          await controller.postContactSearch(modifiedRequest, h);
-          const args = h.redirect.lastCall.args;
-          expect(args[0]).to.equal(`/invoice-accounts/create/${regionId}/${companyId}/select-address`);
-        });
       });
     });
   });
