+++ test/internal/modules/invoice-accounts/controllers/invoice-accounts.js
@@ -4,7 +4,9 @@ const {
   experiment,
   test,
   beforeEach,
-  afterEach
+  afterEach,
+  before,
+  after
 } = exports.lab = require('@hapi/lab').script();
 const controller = require('../../../../../src/internal/modules/invoice-accounts/controllers/invoice-accounts');
 const uuid = require('uuid');
@@ -75,6 +77,7 @@ experiment('./internal/modules/invoice-accounts/controller', () => {
           id: companyId,
           name: companyName
         }],
+        contactSearchResults: [],
         company: {
           id: companyId,
           name: companyName
@@ -430,6 +433,71 @@ experiment('./internal/modules/invoice-accounts/controller', () => {
       const details = [companyId, { address: { addressId }, agent: { companyId: agentId }, contact: undefined, regionId, startDate }];
       const args = dataService.saveInvoiceAccDetails.lastCall.args;
       expect(args).to.equal(details);
+    });
+  });
+
+  experiment('.getSearchCompany', () => {
+    beforeEach(async () => {
+      await dataService.sessionManager.returns(sessionData());
+      await controller.getSearchCompany(request, h);
+    });
+    afterEach(async () => {
+      await sandbox.restore();
+    });
+    test('calls sessionManager with the correct params', async () => {
+      const args = dataService.sessionManager.lastCall.args;
+      expect(args[0]).to.equal(request);
+      expect(args[1]).to.equal(regionId);
+      expect(args[2]).to.equal(companyId);
+      // no data to merge is passed to the session
+      expect(args[3]).to.equal(undefined);
+    });
+
+    test('the correct data is passed to the view', async () => {
+      dataService.sessionManager.returns(sessionData());
+      await controller.getSearchCompany(request, h);
+      const args = h.view.lastCall.args[1];
+      expect(args.back).to.startWith(`/invoice-accounts/create/${regionId}/${companyId}`);
+      expect(args.pageTitle).to.equal('Does this contact already exist?');
+    });
+  });
+
+  experiment('.postSearchCompany', async () => {
+    experiment('when the form is valid', () => {
+      experiment('when the user opts to create a new contact', () => {
+        let modifiedRequest;
+        beforeEach(async () => {
+          modifiedRequest = { payload: { filter: 'some string', id: 'new' } };
+          Object.assign(modifiedRequest, request);
+          await dataService.sessionManager.returns(sessionData());
+          await controller.postSearchCompany(modifiedRequest, h);
+        });
+        afterEach(async () => {
+          await sandbox.restore();
+        });
+        test('then redirect to contact entry', async () => {
+          await controller.postSearchCompany(modifiedRequest, h);
+          const args = h.redirect.lastCall.args;
+          expect(args[0]).to.startWith(`/contact-entry/new`);
+        });
+      });
+      experiment('when the user opts to select an existing contact', () => {
+        const contactId = uuid();
+        let modifiedRequest;
+        before(async () => {
+          modifiedRequest = { payload: { filter: 'some string', id: contactId } };
+          Object.assign(modifiedRequest, request);
+          await controller.postSearchCompany(modifiedRequest, h);
+        });
+        after(async () => {
+          await sandbox.restore();
+        });
+        test('then redirect to select an address', async () => {
+          await controller.postSearchCompany(modifiedRequest, h);
+          const args = h.redirect.lastCall.args;
+          expect(args[0]).to.equal(`/invoice-accounts/create/${regionId}/${companyId}/select-address`);
+        });
+      });
     });
   });
 });
